refactor get geolocation
