+++ src/utils/helpers.ts
@@ -1,4 +1,8 @@
 import moment from 'moment';
+import Geolocation from 'react-native-geolocation-service';
+import Config from 'react-native-config';
+
+import { Location } from '../store/settings/types';
 
 // 60 minutes = 3600 seconds
 const STEP_60 = 3600;
@@ -36,3 +40,46 @@ export const getSliderStepSeconds = (sliderStep: number): number => {
   if (sliderStep === 30) return STEP_30;
   return STEP_15;
 };
+
+const timeSeriesUrl = `https://data.fmi.fi/fmi-apikey/${Config.API_KEY}/timeseries?param=geoid,name,latitude,longitude,region,country&timesteps=2&format=json&attributes=geoid`;
+
+export const getGeolocation = (
+  callback: (arg0: Location, arg1: boolean) => void
+) =>
+  Geolocation.getCurrentPosition(
+    (position) => {
+      const { latitude, longitude } = position.coords;
+      fetch(`${timeSeriesUrl}&latlon=${latitude},${longitude}`)
+        .then((res) => res.json())
+        .then((json) => {
+          const geoid = Number(Object.keys(json)[0]);
+          const vals: {
+            name: string;
+            latitude: number;
+            longitude: number;
+            region: string;
+          }[][] = Object.values(json);
+
+          const { name, region } = vals[0][0];
+          callback(
+            {
+              lat: latitude,
+              lon: longitude,
+              name,
+              area: region,
+              id: geoid,
+            },
+            true
+          );
+        })
+        .catch((e) => console.error(e));
+    },
+    (error) => {
+      console.log('GEOLOCATION NOT AVAILABLE', error);
+    },
+    {
+      enableHighAccuracy: true,
+      timeout: 15000,
+      maximumAge: 10000,
+    }
+  );
